% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_id_visit.R
\name{check_id_visit}
\alias{check_id_visit}
\title{Check ID and visit specification for I2C2}
\usage{
check_id_visit(y, id, visit)
}
\arguments{
\item{y}{An n by p data matrix containing n vectorized image data with p voxels.
Each row contains one observed image data at a particular visit for one subject.
Each column contains image values for all subjects and visits at a particular voxel.

The rows are organized by subjects and then visits, EX)
(Y11, Y12, Y21, Y22, ... , YI1 , YI2)}

\item{id}{Vector of IDs, EX) c(1, 1, 2, 2, 3, 3, 4, 4, ... , I, I)}

\item{visit}{Vector of visits, EX) (1, 2, 1, 2, 1, 2, ... , 1, 2)}
}
\value{
List of elements: reordered data, id, visit, I (number of ids),
number of rows (n), and number of columns (p)
}
\description{
Checks the input for I2C2
}
\examples{
id = c(1:10, 10:1)
visit = rep(1:2, each = 10)
visit = as.character(visit)
n = length(id)
p = 100
y = matrix(rnorm(n * p), nrow = n, ncol = p)
check_id_visit(y =y, id = id, visit = visit)
}
