% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demean_matrix.R
\name{demean_matrix}
\alias{demean_matrix}
\title{Worker function for demeaning a matrix}
\usage{
demean_matrix(y, visit, twoway = TRUE, tol = 0)
}
\arguments{
\item{y}{n by p matrix of data}

\item{visit}{Vector of visits, EX) (1, 2, 1, 2, 1, 2, ... , 1, 2)}

\item{twoway}{a logical argument indicating whether a oneway or twoway
mean subtraction is more appropriate for the problem. If FALSE, only the overall sample
mean will be removed only; if TRUE, it will also remove visit specific means to
avoid scanner or batch effects}

\item{tol}{tolerance for matching visits.}
}
\value{
Matrix of the same size as y
}
\description{
De-means a matrix based on the column means or the
visit-specific means.
}
\examples{
id = c(1:10, 1:10)
visit = rep(1:2, each = 10)
n = length(id)
p = 100
y = matrix(rnorm(n * p), nrow = n, ncol = p)
demean_matrix(y, visit = visit)
}
