% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I2C2.mcCI.R
\name{I2C2.mcCI}
\alias{I2C2.mcCI}
\title{Compute the Convidence Interval}
\usage{
I2C2.mcCI(..., rseed = 1234, R = 100, mc.cores = 1, ci = 0.95)
}
\arguments{
\item{...}{arguments passed to \code{\link{I2C2.rsample}}}

\item{rseed}{Seed number}

\item{R}{The bootstrap repetition size}

\item{mc.cores}{Number of Cores}

\item{ci}{100*ci\% The level of the Confidence Interval}
}
\value{
List of the lambdas and then the confidence interval
}
\description{
Computing the confidence interval of I2C2 via multicore computing.
}
