% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I2C2.mcNulldist.R
\name{I2C2.mcNulldist}
\alias{I2C2.mcNulldist}
\title{Compute the Null Distribution}
\usage{
I2C2.mcNulldist(..., rseed = 1234, R = 500, mc.cores = 1)
}
\arguments{
\item{...}{arguments passed to \code{\link{I2C2.rpermute}}}

\item{rseed}{Seed number}

\item{R}{Number of permutations}

\item{mc.cores}{Number of Cores}
}
\value{
List of the lambdas
}
\description{
Computing the null distribution of I2C2
via permutation and multicore computing.
}
