% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i2c2.R
\name{I2C2}
\alias{I2C2}
\alias{i2c2}
\title{Image Intraclass Correlation Coefficient}
\usage{
I2C2(y, id, visit, symmetric = FALSE, truncate = FALSE, twoway = TRUE,
  demean = TRUE, ...)

i2c2(...)
}
\arguments{
\item{y}{An n by p data matrix containing functional responses.
Each row contains measurements from a function for one observation at a
set of grid points, and each column contains measurements of all
functions at a particular grid point.

The rows are organized by subjects and then visits, EX)
(Y11, Y12, Y21, Y22, ... , YI1 , YI2)}

\item{id}{Vector of IDs, EX) c(1, 1, 2, 2, 3, 3, 4, 4, ... , I, I)}

\item{visit}{Vector of visits, EX) (1, 2, 1, 2, 1, 2, ... , 1, 2)}

\item{symmetric}{if FALSE then the function uses the
method of moments estimator formula;
if TRUE, pairwise symmetric sum formula, default is FALSE}

\item{truncate}{if TRUE, set negative I2C2 to zero}

\item{twoway}{a logical argument indicating whether a oneway or a twoway
functional ANOVA (analysis of variance) decomposition is more
appropriate for the problem. "twoway = TRUE" will carry out twoway
ANOVA and remove the visit specific mean}

\item{demean}{if TRUE, include the demean step and
output the demeaned dataset}

\item{...}{not used}
}
\value{
The output of the function is a list that contains the
following elements.
lambda:       estimated I2C2
Kx:           the trace of between-cluster variance operator
Ku:           the trace of within-cluster variance operator
demean_y:     if demean == TRUE, output the demeaned dataset
}
\description{
Calculate image intraclass correlation
 coefficient (I2C2) of balanced/unbalanced data using the trace method
}
\examples{
id = c(1:10, 10:1)
visit = rep(1:2, each = 10)
visit = as.character(visit)
n = length(id)
p = 100
y = matrix(rnorm(n * p), nrow = n, ncol = p)
res = I2C2(y, id = id, visit = visit)
res$lambda
}
\references{
Haochang Shou, Ani Eloyan, Seonjoo Lee, Vadim Zipunnikov, Mary Beth Nebel,
  Brian Caffo, Martin Lindquist, Ciprian M. Crainiceanu  (2012) The image intra
class correlation coefficient for replication studies.
}
\author{
Haochang Shou, Ani Eloyan, Seonjoo Lee, Vadim Zipunnikov,
Mary Beth Nebel, Brian Caffo, Martin Lindquist, Ciprian M. Crainiceanu
}
