% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I2C2_rpermute.R
\name{I2C2.rpermute}
\alias{I2C2.rpermute}
\title{Compute I2C2 for Permuted Data}
\usage{
I2C2.rpermute(s = 1, y, id, visit, ...)
}
\arguments{
\item{s}{Random seed, Default = 1}

\item{...}{arguments to pass to \code{\link{I2C2}}}
}
\value{
Set of lambda values

Lambda value from \code{\link{I2C2}}
}
\description{
Computing the I2C2 of the permuted data.
}
