% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I2C2.rsample.R
\name{I2C2.rsample}
\alias{I2C2.rsample}
\alias{resample_id}
\title{Compute I2C2 for Permuted Data}
\usage{
I2C2.rsample(s = 1, y, id, visit, ...)

resample_id(s = 1, y, id, visit)
}
\arguments{
\item{s}{Random seed, Default = 1}

\item{id}{Vector of IDs, EX) c(1, 1, 2, 2, 3, 3, 4, 4, ... , I, I)}

\item{visit}{Vector of visits, EX) (1, 2, 1, 2, 1, 2, ... , 1, 2)}

\item{...}{arguments to pass to \code{\link{I2C2}}}
}
\value{
Set of lambda values
}
\description{
Computing the I2C2 of the permuted data.
}
\examples{
id = c(1:10, 10:1)
visit = rep(1:2, each = 10)
visit = as.character(visit)
n = length(id)
p = 100
y = matrix(rnorm(n * p), nrow = n, ncol = p)
L = resample_id(s = 10, y = y, id = id, visit = visit)
}
