% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask10.R
\docType{data}
\name{mask10}
\alias{mask10}
\title{Mask for Data in Image}
\format{An array of 256 by 256 by 181 voxels indicating the \code{11294}
voxels in the mask}
\usage{
mask10
}
\description{
A matrix of voxels in the ventricle mask
}
\keyword{datasets}
