% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_large_matrix.R
\name{print.demeaned_matrix}
\alias{print.demeaned_matrix}
\title{Print method for `demeaned_matrix`}
\usage{
\method{print}{demeaned_matrix}(x, reveal = FALSE, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{reveal}{Should the `demeaned_matrix` be revealed}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print method for `demeaned_matrix`
}
\examples{
x = I2C2::mask10
class(x) = "demeaned_matrix"
print(x)
print(x, reveal = TRUE)
}
