% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vn10.R
\docType{data}
\name{vn10}
\alias{vn10}
\title{Example RAVENS data from the Ventricles}
\format{A list with 3 elements, which are:
\describe{
\item{y}{matrix of values for 42 id/visit combinations and the 11294
voxels in the ventricles}
\item{id}{vector of ids}
\item{visit}{vector of visit indicators}
}}
\usage{
vn10
}
\description{
A list containing the data from the ventricles from
RAVENS maps
}
\keyword{datasets}
