% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I2C2_rpermute.R
\name{I2C2.rpermute}
\alias{I2C2.rpermute}
\title{Compute I2C2 for Permuted Data}
\usage{
I2C2.rpermute(s = 1, y, id, visit, ...)
}
\arguments{
\item{s}{Random seed, Default = 1}

\item{y}{An n by p data matrix containing functional responses.
Each row contains measurements from a function for one observation at a
set of grid points, and each column contains measurements of all
functions at a particular grid point.

The rows are organized by subjects and then visits, EX)
(Y11, Y12, Y21, Y22, ... , YI1 , YI2)}

\item{id}{Vector of IDs, EX) c(1, 1, 2, 2, 3, 3, 4, 4, ... , I, I)}

\item{visit}{Vector of visits, EX) (1, 2, 1, 2, 1, 2, ... , 1, 2)}

\item{...}{arguments to pass to \code{\link{I2C2}}}
}
\value{
Set of lambda values

Lambda value from \code{\link{I2C2}}
}
\description{
Computing the I2C2 of the permuted data.
}
